refactor: UI와 연관된 message 상수를 View로 이동 및 사용자 입력은 query로 분리
