+++ src/View/Constants/message.js
@@ -1,0 +1,6 @@
+export const RACE = {
+  RESULT_INSTRUCTION: '실행 결과',
+  WINNER_INSTRUCTION: '최종 우승자:',
+  WINNER_DELIMITER: ', ',
+  MOVEMENT: '-',
+};
