+++ src/View/Constants/query.js
@@ -1,10 +1,3 @@
-export const RACE = {
-  RESULT_INSTRUCTION: '실행 결과',
-  WINNER_INSTRUCTION: '최종 우승자:',
-  WINNER_DELIMITER: ', ',
-  MOVEMENT: '-',
-};
-
 export const INPUT = {
   CARS_NAME: '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).',
   GAME_COUNT: '시도할 횟수는 몇 회인가요?',
